Noosgi/standalone (#541)
* Database initialisation overhaul with Liquibase

As those databases don't have sequence, add autoincrement for mysql and mariadb.
For info, Maria db now support create sequence

* Update all Id type for model from int to long (#534)

Id became BIGINT into database
Use only primitive type UNLESS is nullable according to the database schema (Liquibase).
Update CI to test only on LTS database

* Add standalone version of the simple WS (#532)

* Add kubernetes template for standalone deployment (#538)

* Merge kube-specific deploy files with docker ones

---------

+++ jqm-all/jqm-admin/src/main/java/com/enioka/api/admin/NodeDto.java
@@ -25,7 +25,7 @@ public class NodeDto implements Serializable
 {
     private static final long serialVersionUID = -3592156944832541035L;
 
-    private Integer id;
+    private Long id;
     private String name;
     private String dns;
     private Integer port;
@@ -39,12 +39,12 @@ public class NodeDto implements Serializable
     private Boolean loapApiSimple, loadApiClient, loadApiAdmin;
     private Boolean reportsRunning;
 
-    public Integer getId()
+    public Long getId()
     {
         return id;
     }
 
-    public void setId(Integer id)
+    public void setId(Long id)
     {
         this.id = id;
     }
