+++ jqm-all/jqm-engine/src/main/java/com/enioka/jqm/engine/RunnerManager.java
@@ -21,7 +21,7 @@ class RunnerManager implements AutoCloseable
 {
     private Logger jqmlogger = LoggerFactory.getLogger(RunnerManager.class);
 
-    private Map<Integer, JobRunner> runnerCache = new HashMap<>();
+    private Map<Long, JobRunner> runnerCache = new HashMap<>();
 
     private ConcurrentLinkedQueue<JobRunner> runners = new ConcurrentLinkedQueue<>();
