+++ jqm-all/jqm-integration-tests/src/test/java/com/enioka/jqm/integration/tests/CliTest.java
@@ -64,7 +64,7 @@ public class CliTest extends JqmBaseTest
     {
         CreationTools.createJobDef(null, true, "App", null, "jqm-tests/jqm-test-datetimemaven/target/test.jar", TestHelpers.qVip, 42,
                 "MarsuApplication", null, "Franquin", "ModuleMachin", "other", "other", true, cnx);
-        int i = jqmClient.newJobRequest("MarsuApplication", "TestUser").enqueue();
+        Long i = jqmClient.newJobRequest("MarsuApplication", "TestUser").enqueue();
         cnx.runUpdate("ji_update_status_by_id", TestHelpers.node.getId(), i);
         cnx.runUpdate("debug_jj_update_node_by_id", TestHelpers.node.getId(), i);
         cnx.commit();
