+++ jqm-all/jqm-tst/src/main/java/com/enioka/jqm/test/api/JqmAsynchronousTester.java
@@ -149,7 +149,7 @@ public interface JqmAsynchronousTester extends AutoCloseable
      *
      * @return the request ID.
      */
-    public int enqueue(String name);
+    public Long enqueue(String name);
 
     /**
      * Wait for a given amount of ended job instances (OK or KO).
